web/requirements.txt
Django==1.6
PyVirtualDisplay==0.1.5
argparse==1.2.1
boto==2.20.1
django-storages==1.1.8
jsonpatch==1.3
jsonpickle==0.8.0
jsonpointer==1.0
oauth==1.0.1
requests==2.2.1
schedule==0.3.1
selenium==2.44.0
splinter==0.7.0
pyPdf==1.13
django-analytical==0.19.0
MySQL-python==1.2.5